refactor: port level scheduling
