--- ognisko/constants/level_comments.py
@@ -1,9 +1,0 @@
-from __future__ import annotations
-
-from enum import IntEnum
-
-
-# NOTE: If used anywhere else, move to a more generic location.
-class LevelCommentSorting(IntEnum):
-    NEWEST = 0
-    MOST_LIKED = 1
