--- ognisko/constants/level_schedules.py
@@ -1,8 +1,0 @@
-from __future__ import annotations
-
-from enum import IntEnum
-
-
-class LevelScheduleType(IntEnum):
-    DAILY = 0
-    WEEKLY = 1
