--- ognisko/models/__init__.py
@@ -1,14 +1,0 @@
-from __future__ import annotations
-
-from . import daily_chest
-from . import friend_request
-from . import level
-from . import level_comment
-from . import level_schedule
-from . import like
-from . import rgb
-from . import song
-from . import user
-from . import user_comment
-from . import user_credential
-from . import user_relationship
